Some decoupling, and various refactor efforts. Adding logic to avoid adding TEC hooks unless import is started, likely a bug if left always bound.
